requirements.txt
agno
python-dotenv
anthropic
